M2-PR-04: convert integration_tests.rs to directory + extract common/
Behaviour-preserving. 7176 LOC flat file becomes:
- integration_tests/common/polling.rs (90): POLL_*/LONG_TEST_TIMEOUT
  constants, 4 env locks, EnvVarGuard, wait_for, wait_for_async,
  with_*_test_timeout.
- integration_tests/common/mock_kanata.rs (191): KanataMessage,
  wait_for/drain_kanata_messages, MockKanataConfig, MockKanataServer.
- integration_tests/common/focus_service.rs (149):
  start_wayland_test_server, pause/unpause_daemon_direct,
  FocusService zbus interface, TEST_DAEMON_DBUS_NAME,
  start_gnome_focus_service.
- integration_tests/common/dbus_session.rs (119):
  dbus_daemon_available, DBUS_TEST_COUNTER, DbusSessionGuard.
- integration_tests/common/mod.rs (11): mod decls + pub(crate) use.
- integration_tests/mod.rs (6638): all 65 tests + mod common.

src/daemon/integration_tests.rs deleted.

One zbus visibility tweak: FocusService::focus_changed signal made
pub (was implicit-private when the interface lived next to callers).
No other production source changes.

start_wayland_test_server stays in common/focus_service.rs;
references super::super::wayland_mock::WaylandMockServer
(inline child still in mod.rs). No cycle.

Verification: 65 test attrs preserved; cargo build clean;
cargo test --bin kanata-switcher 261 passed / 0 failed.
