+++ src/daemon/integration_tests/mod.rs
@@ -17,428 +17,9 @@ use std::sync::mpsc;
 use std::thread;
 use std::time::{Duration, Instant};
 
-// === Polling Helper ===
 
-const POLL_INTERVAL: Duration = Duration::from_millis(50);
-const POLL_TIMEOUT: Duration = Duration::from_secs(5);
-const TEST_TIMEOUT: Duration = Duration::from_secs(5);
-const LONG_TEST_TIMEOUT: Duration = Duration::from_secs(20);
-static WAYLAND_ENV_LOCK: std::sync::Mutex<()> = std::sync::Mutex::new(());
-static DBUS_ENV_LOCK: std::sync::Mutex<()> = std::sync::Mutex::new(());
-static X11_FOCUS_QUERY_LOCK: std::sync::Mutex<()> = std::sync::Mutex::new(());
-static DISPLAY_ENV_LOCK: std::sync::Mutex<()> = std::sync::Mutex::new(());
-
-struct EnvVarGuard {
-    key: &'static str,
-    previous: Option<std::ffi::OsString>,
-}
-
-impl EnvVarGuard {
-    fn set(key: &'static str, value: &str) -> Self {
-        let previous = std::env::var_os(key);
-        unsafe {
-            std::env::set_var(key, value);
-        }
-        Self { key, previous }
-    }
-}
-
-impl Drop for EnvVarGuard {
-    fn drop(&mut self) {
-        match self.previous.take() {
-            Some(value) => unsafe {
-                std::env::set_var(self.key, value);
-            },
-            None => unsafe {
-                std::env::remove_var(self.key);
-            },
-        }
-    }
-}
-
-/// Wait for a condition to become true, polling at 50ms intervals.
-/// Returns Ok(T) when the condition returns Some(T), or Err after 1 minute timeout.
-fn wait_for<T, F>(mut condition: F) -> Result<T, &'static str>
-where
-    F: FnMut() -> Option<T>,
-{
-    let start = Instant::now();
-    while start.elapsed() < POLL_TIMEOUT {
-        if let Some(result) = condition() {
-            return Ok(result);
-        }
-        thread::sleep(POLL_INTERVAL);
-    }
-    Err("Timeout waiting for condition")
-}
-
-/// Async version of wait_for for tokio tests
-async fn wait_for_async<T, F, Fut>(mut condition: F) -> Result<T, &'static str>
-where
-    F: FnMut() -> Fut,
-    Fut: std::future::Future<Output = Option<T>>,
-{
-    let start = Instant::now();
-    while start.elapsed() < POLL_TIMEOUT {
-        if let Some(result) = condition().await {
-            return Ok(result);
-        }
-        tokio::time::sleep(POLL_INTERVAL).await;
-    }
-    Err("Timeout waiting for condition")
-}
-
-async fn with_test_timeout<F, T>(future: F) -> T
-where
-    F: Future<Output = T>,
-{
-    tokio::time::timeout(TEST_TIMEOUT, future)
-        .await
-        .expect("test timeout")
-}
-
-async fn with_long_test_timeout<F, T>(future: F) -> T
-where
-    F: Future<Output = T>,
-{
-    tokio::time::timeout(LONG_TEST_TIMEOUT, future)
-        .await
-        .expect("test timeout")
-}
-
-fn start_wayland_test_server() -> (
-    std::sync::MutexGuard<'static, ()>,
-    wayland_mock::WaylandMockServer,
-) {
-    let lock = WAYLAND_ENV_LOCK.lock().unwrap();
-    let server = wayland_mock::WaylandMockServer::start();
-    (lock, server)
-}
-
-async fn pause_daemon_direct(
-    pause_broadcaster: &PauseBroadcaster,
-    handler: &Arc<Mutex<FocusHandler>>,
-    status_broadcaster: &StatusBroadcaster,
-    kanata: &KanataClient,
-    request_label: &str,
-) {
-    if !pause_broadcaster.set_paused(true) {
-        println!("[Pause] Pause requested {} (already paused)", request_label);
-        return;
-    }
-    println!("[Pause] Pausing daemon");
-    let virtual_keys = {
-        let mut handler = handler.lock().unwrap();
-        let keys = handler.current_virtual_keys();
-        handler.reset();
-        keys
-    };
-    let default_layer = kanata.default_layer().await.unwrap_or_default();
-
-    for vk in virtual_keys.iter().rev() {
-        kanata.act_on_fake_key(vk, "Release").await;
-    }
-
-    if !default_layer.is_empty() {
-        let _ = kanata.change_layer(&default_layer).await;
-    }
-
-    status_broadcaster.set_paused_status(default_layer);
-    kanata.pause_disconnect().await;
-}
-
-async fn unpause_daemon_direct(
-    env: Environment,
-    connection: Option<Connection>,
-    is_kde6: bool,
-    pause_broadcaster: &PauseBroadcaster,
-    handler: &Arc<Mutex<FocusHandler>>,
-    status_broadcaster: &StatusBroadcaster,
-    kanata: &KanataClient,
-    request_label: &str,
-) {
-    if !pause_broadcaster.set_paused(false) {
-        println!(
-            "[Pause] Unpause requested {} (already running)",
-            request_label
-        );
-        return;
-    }
-    println!("[Pause] Resuming daemon");
-    kanata.unpause_connect().await;
-    if let Err(error) = apply_focus_for_env(
-        env,
-        connection.as_ref(),
-        is_kde6,
-        handler,
-        status_broadcaster,
-        pause_broadcaster,
-        kanata,
-    )
-    .await
-    {
-        panic!("[Pause] Failed to refresh focus after unpause: {}", error);
-    }
-}
-
-// === Mock Kanata Server ===
-
-/// Messages that can be sent to the mock Kanata server
-#[derive(Debug, Clone, PartialEq)]
-enum KanataMessage {
-    ChangeLayer { new: String },
-    ActOnFakeKey { name: String, action: String },
-    RequestLayerNames,
-    RequestFakeKeyNames,
-}
-
-struct FocusService {
-    call_count: Arc<std::sync::atomic::AtomicUsize>,
-    class: String,
-    title: String,
-}
-
-#[zbus::interface(name = "com.github.kanata.Switcher.extensions.GNOME")]
-impl FocusService {
-    #[allow(non_snake_case)]
-    fn GetFocus(&self) -> (String, String) {
-        self.call_count
-            .fetch_add(1, std::sync::atomic::Ordering::SeqCst);
-        (self.class.clone(), self.title.clone())
-    }
-
-    #[zbus(signal, name = "FocusChanged")]
-    async fn focus_changed(
-        signal_emitter: &zbus::object_server::SignalEmitter<'_>,
-        class: &str,
-        title: &str,
-    ) -> zbus::Result<()>;
-}
-
-/// Bus name used by integration tests when registering the daemon's control
-/// service on a private session bus. All test sites that previously hardcoded
-/// `com.github.kanata.Switcher` should use this constant so the per-instance
-/// name plumbing is exercised consistently.
-const TEST_DAEMON_DBUS_NAME: &str = "com.github.kanata.Switcher.instances.test";
-
-async fn start_gnome_focus_service(
-    address: &zbus::Address,
-    class: &str,
-    title: &str,
-) -> (Connection, Arc<std::sync::atomic::AtomicUsize>) {
-    use zbus::connection::Builder;
-
-    let call_count = Arc::new(std::sync::atomic::AtomicUsize::new(0));
-    let service_connection = Builder::address(address.clone())
-        .expect("Failed to create connection builder")
-        .name(GNOME_SHELL_BUS_NAME)
-        .expect("Failed to set bus name")
-        .serve_at(
-            GNOME_FOCUS_OBJECT_PATH,
-            FocusService {
-                call_count: call_count.clone(),
-                class: class.to_string(),
-                title: title.to_string(),
-            },
-        )
-        .expect("Failed to serve mock focus service")
-        .build()
-        .await
-        .expect("Failed to build focus service connection");
-
-    let dbus_proxy = zbus::fdo::DBusProxy::new(&service_connection)
-        .await
-        .expect("Failed to create DBus proxy");
-    wait_for_async(|| {
-        let proxy = dbus_proxy.clone();
-        async move {
-            proxy
-                .name_has_owner(GNOME_SHELL_BUS_NAME.try_into().unwrap())
-                .await
-                .ok()
-                .filter(|&has_owner| has_owner)
-        }
-    })
-    .await
-    .expect("Timeout waiting for GNOME focus service registration");
-
-    (service_connection, call_count)
-}
-
-fn wait_for_kanata_message(
-    server: &MockKanataServer,
-    message: KanataMessage,
-    timeout_duration: Duration,
-) {
-    let start = Instant::now();
-    while start.elapsed() < timeout_duration {
-        if let Some(msg) = server.recv_timeout(Duration::from_millis(50)) {
-            if msg == message {
-                return;
-            }
-        }
-    }
-    panic!("Timeout waiting for {:?}", message);
-}
-
-fn drain_kanata_messages(server: &MockKanataServer, duration: Duration) {
-    let start = Instant::now();
-    while start.elapsed() < duration {
-        if server.recv_timeout(Duration::from_millis(20)).is_none() {
-            break;
-        }
-    }
-}
-/// Configuration for MockKanataServer
-struct MockKanataConfig {
-    /// Virtual keys to report. If None, simulate older kanata that doesn't support the command.
-    virtual_keys: Option<Vec<String>>,
-}
-
-impl Default for MockKanataConfig {
-    fn default() -> Self {
-        Self {
-            virtual_keys: Some(vec![
-                "vk_browser".to_string(),
-                "vk_terminal".to_string(),
-                "vk_vim".to_string(),
-            ]),
-        }
-    }
-}
-
-/// A mock Kanata TCP server for testing
-struct MockKanataServer {
-    port: u16,
-    handle: Option<thread::JoinHandle<()>>,
-    receiver: mpsc::Receiver<KanataMessage>,
-    shutdown: std::sync::Arc<std::sync::atomic::AtomicBool>,
-}
-
-impl MockKanataServer {
-    fn start() -> Self {
-        Self::start_with_config(MockKanataConfig::default())
-    }
-
-    /// Start a mock server simulating older kanata that doesn't support RequestFakeKeyNames
-    fn start_legacy() -> Self {
-        Self::start_with_config(MockKanataConfig { virtual_keys: None })
-    }
-
-    fn start_with_config(config: MockKanataConfig) -> Self {
-        let listener = TcpListener::bind("127.0.0.1:0").unwrap();
-        listener.set_nonblocking(true).unwrap();
-        let port = listener.local_addr().unwrap().port();
-        let (sender, receiver) = mpsc::channel();
-        let shutdown = std::sync::Arc::new(std::sync::atomic::AtomicBool::new(false));
-        let shutdown_thread = std::sync::Arc::clone(&shutdown);
-
-        let handle = thread::spawn(move || {
-            loop {
-                if shutdown_thread.load(std::sync::atomic::Ordering::SeqCst) {
-                    break;
-                }
-                let (mut stream, _) = match listener.accept() {
-                    Ok(connection) => connection,
-                    Err(err) if err.kind() == std::io::ErrorKind::WouldBlock => {
-                        thread::sleep(Duration::from_millis(50));
-                        continue;
-                    }
-                    Err(_) => break,
-                };
-                stream.set_read_timeout(Some(Duration::from_secs(5))).ok();
-
-                // Send initial LayerChange message
-                let init_msg = r#"{"LayerChange":{"new":"default"}}"#;
-                if writeln!(stream, "{}", init_msg).is_err() {
-                    continue;
-                }
-
-                let mut reader = BufReader::new(stream.try_clone().unwrap());
-
-                loop {
-                    let mut line = String::new();
-                    match reader.read_line(&mut line) {
-                        Ok(0) => break, // Connection closed
-                        Ok(_) => {
-                            // Parse and forward the message
-                            if let Ok(value) = serde_json::from_str::<serde_json::Value>(&line) {
-                                if let Some(cl) = value.get("ChangeLayer") {
-                                    let new = cl.get("new").and_then(|v| v.as_str()).unwrap_or("");
-                                    sender
-                                        .send(KanataMessage::ChangeLayer {
-                                            new: new.to_string(),
-                                        })
-                                        .ok();
-                                } else if let Some(fk) = value.get("ActOnFakeKey") {
-                                    let name =
-                                        fk.get("name").and_then(|v| v.as_str()).unwrap_or("");
-                                    let action =
-                                        fk.get("action").and_then(|v| v.as_str()).unwrap_or("");
-                                    sender
-                                        .send(KanataMessage::ActOnFakeKey {
-                                            name: name.to_string(),
-                                            action: action.to_string(),
-                                        })
-                                        .ok();
-                                } else if value.get("RequestLayerNames").is_some() {
-                                    sender.send(KanataMessage::RequestLayerNames).ok();
-                                    // Respond with layer names
-                                    let response = r#"{"LayerNames":{"names":["default","browser","terminal","vim"]}}"#;
-                                    writeln!(stream, "{}", response).ok();
-                                } else if value.get("RequestFakeKeyNames").is_some() {
-                                    sender.send(KanataMessage::RequestFakeKeyNames).ok();
-                                    // Respond based on config
-                                    match &config.virtual_keys {
-                                        Some(vks) => {
-                                            let names_json = serde_json::to_string(vks).unwrap();
-                                            let response = format!(
-                                                r#"{{"FakeKeyNames":{{"names":{}}}}}"#,
-                                                names_json
-                                            );
-                                            writeln!(stream, "{}", response).ok();
-                                        }
-                                        None => {
-                                            // Simulate older kanata: send error then drop connection
-                                            let response = r#"{"status":"Error","msg":"Failed to deserialize command: unknown variant `RequestFakeKeyNames`"}"#;
-                                            writeln!(stream, "{}", response).ok();
-                                            break;
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                        Err(_) => break,
-                    }
-                }
-            }
-        });
-        Self {
-            port,
-            handle: Some(handle),
-            receiver,
-            shutdown,
-        }
-    }
-
-    fn port(&self) -> u16 {
-        self.port
-    }
-
-    fn recv_timeout(&self, timeout: Duration) -> Option<KanataMessage> {
-        self.receiver.recv_timeout(timeout).ok()
-    }
-}
-
-impl Drop for MockKanataServer {
-    fn drop(&mut self) {
-        self.shutdown
-            .store(true, std::sync::atomic::Ordering::SeqCst);
-        if let Some(handle) = self.handle.take() {
-            let _ = handle.join();
-        }
-    }
-}
+mod common;
+pub(super) use common::*;
 
 // === GNOME Focus Query Tests ===
 
@@ -1567,125 +1148,6 @@ async fn test_dbus_service_fallthrough() {
     .await;
 }
 
-// === Private DBus Session for Testing ===
-
-/// Check if dbus-daemon is available by trying to run it with --version
-fn dbus_daemon_available() -> bool {
-    std::process::Command::new("dbus-daemon")
-        .arg("--version")
-        .output()
-        .map(|o| o.status.success())
-        .unwrap_or(false)
-}
-
-/// Guard struct that starts a private dbus-daemon and cleans up on drop
-struct DbusSessionGuard {
-    child: std::process::Child,
-    address: String,
-    config_dir: std::path::PathBuf,
-}
-
-use std::sync::atomic::{AtomicU64, Ordering};
-static DBUS_TEST_COUNTER: AtomicU64 = AtomicU64::new(0);
-
-impl DbusSessionGuard {
-    fn start() -> Result<Self, String> {
-        if !dbus_daemon_available() {
-            return Err("dbus-daemon binary not found in PATH".to_string());
-        }
-
-        // Create a minimal session config file with unique path per test
-        let unique_id = DBUS_TEST_COUNTER.fetch_add(1, Ordering::SeqCst);
-        let config_dir =
-            std::env::temp_dir().join(format!("dbus-test-{}-{}", std::process::id(), unique_id));
-        std::fs::create_dir_all(&config_dir)
-            .map_err(|e| format!("Failed to create config dir: {}", e))?;
-
-        let config_path = config_dir.join("session.conf");
-        let socket_path = config_dir.join("bus-socket");
-
-        // Minimal session bus config
-        let config_content = format!(
-            r#"<!DOCTYPE busconfig PUBLIC "-//freedesktop//DTD D-Bus Bus Configuration 1.0//EN" "http://www.freedesktop.org/standards/dbus/1.0/busconfig.dtd">
-<busconfig>
-  <type>session</type>
-  <listen>unix:path={}</listen>
-  <policy context="default">
-    <allow send_destination="*" eavesdrop="true"/>
-    <allow eavesdrop="true"/>
-    <allow own="*"/>
-  </policy>
-</busconfig>"#,
-            socket_path.display()
-        );
-
-        std::fs::write(&config_path, config_content)
-            .map_err(|e| format!("Failed to write config file: {}", e))?;
-
-        // Start dbus-daemon with custom config
-        let mut child = std::process::Command::new("dbus-daemon")
-            .args([
-                "--config-file",
-                config_path.to_str().unwrap(),
-                "--nofork",
-                "--print-address",
-            ])
-            .stdout(std::process::Stdio::piped())
-            .stderr(std::process::Stdio::piped())
-            .spawn()
-            .map_err(|e| format!("Failed to spawn dbus-daemon: {}", e))?;
-
-        // Read the address from stdout
-        let stdout = child
-            .stdout
-            .take()
-            .ok_or("Failed to capture dbus-daemon stdout")?;
-        let mut reader = std::io::BufReader::new(stdout);
-        let mut address = String::new();
-        std::io::BufRead::read_line(&mut reader, &mut address)
-            .map_err(|e| format!("Failed to read dbus-daemon address: {}", e))?;
-        let address = address.trim().to_string();
-
-        if address.is_empty() {
-            // Try to read stderr for error info
-            if let Some(mut stderr) = child.stderr.take() {
-                let mut err_output = String::new();
-                let _ = std::io::Read::read_to_string(&mut stderr, &mut err_output);
-                let _ = child.kill();
-                return Err(format!(
-                    "dbus-daemon produced no address. stderr: {}",
-                    err_output
-                ));
-            }
-            let _ = child.kill();
-            return Err("dbus-daemon produced no address".to_string());
-        }
-
-        // Wait for socket to be connectable (dbus-daemon ready)
-        let socket_path_clone = socket_path.clone();
-        wait_for(|| std::os::unix::net::UnixStream::connect(&socket_path_clone).ok())
-            .map_err(|_| "Timeout waiting for dbus-daemon socket")?;
-
-        Ok(Self {
-            child,
-            address,
-            config_dir,
-        })
-    }
-
-    fn address(&self) -> &str {
-        &self.address
-    }
-}
-
-impl Drop for DbusSessionGuard {
-    fn drop(&mut self) {
-        let _ = self.child.kill();
-        let _ = self.child.wait();
-        // Clean up config directory
-        let _ = std::fs::remove_dir_all(&self.config_dir);
-    }
-}
 
 /// Test with a private DBus session (no dependency on desktop session)
 ///
